% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asympt_smirnov}
\alias{asympt_smirnov}
\title{Asymptotic p values in Smirnov distribution.}
\usage{
asympt_smirnov(q, n1, n2, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{q}{quantile of the distribution.}

\item{n1}{number of complete observations in the first sample.}

\item{n2}{number of complete observations in the second sample.}

\item{alternative}{type of alternative hypothesis}
}
\value{
a single numeric value: p value for the given quantile öf
Smirnov distribution.
}
\description{
An internal, simplified wrapper function around \code{\link{psmirnov}}.
}
