% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{chunk}
\alias{chunk}
\alias{chunk.default}
\alias{chunk.list}
\alias{chunk.matrix}
\alias{chunk.data.frame}
\title{Split an object into chunks.}
\usage{
chunk(x, ...)

\method{chunk}{default}(x, n, ...)

\method{chunk}{list}(x, n, ...)

\method{chunk}{matrix}(x, n, ...)

\method{chunk}{data.frame}(x, n, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}

\item{n}{an integer defining the chunk number.}
}
\value{
a list of objects split by elements or columns.
}
\description{
Splits an object (vector, list, matrix or a data frame) into \code{n} chunks.
If the object length or its column number is not divisible by \code{n}, the
output accomplished by \code{cut} is intended to comprise elements of a similar
lengths.
}
