% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_tests.R
\name{f_chisq_test}
\alias{f_chisq_test}
\alias{f_chisq_test.table}
\alias{f_chisq_test.default}
\alias{f_chisq_test.matrix}
\alias{f_chisq_test.data.frame}
\title{Perason's chi-square test.}
\usage{
f_chisq_test(x, ...)

\method{f_chisq_test}{table}(x, correct = TRUE, as_data_frame = FALSE, ...)

\method{f_chisq_test}{default}(x, f, correct = TRUE, as_data_frame = FALSE, ...)

\method{f_chisq_test}{matrix}(
  x,
  f,
  correct = TRUE,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_chisq_test}{data.frame}(
  x,
  f,
  correct = TRUE,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a table object, vector, matrix or a data frame. Vectors, matrices,
and data frames have to contain numeric values, factors, or data compatible
with numeric conversion.}

\item{...}{extra arguments passed to methods.}

\item{correct}{should Yates correction be applied?}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{f}{a factor or integer vector used for splitting.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
a vector/matrix or a data frame with numbers of complete observations,
chi-square statistic, degrees of freedom, p values, Cramer's V
effect size statistics, and, optionally, multiple testing-adjusted p value.
}
\description{
A faster and more versatile alternative to base R's
\code{\link[stats]{chisq.test}}.
}
\details{
In case of matrices or data frames, a series of chi-square tests is performed
in a column-wise manner, i.e. each column is treated as a separate variable.
}
