% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_tests.R
\name{f_cor_test}
\alias{f_cor_test}
\alias{f_cor_test.default}
\alias{f_cor_test.list}
\alias{f_cor_test.matrix}
\alias{f_cor_test.data.frame}
\title{Permutation and bootstrap tests for correlations.}
\usage{
f_cor_test(x, ...)

\method{f_cor_test}{default}(
  x,
  y,
  type = c("permutation", "bootstrap"),
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  ...
)

\method{f_cor_test}{list}(
  x,
  type = c("permutation", "bootstrap"),
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  ...
)

\method{f_cor_test}{matrix}(
  x,
  type = c("permutation", "bootstrap"),
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  adj_method = "none",
  ...
)

\method{f_cor_test}{data.frame}(
  x,
  type = c("permutation", "bootstrap"),
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  adj_method = "none",
  ...
)
}
\arguments{
\item{x}{a numeric vector, list, or data frame.}

\item{...}{extra arguments passed to methods.}

\item{y}{a numeric vector.}

\item{type}{type of the test: permutation (default) or bootstrap.}

\item{method}{type of covariance or correlation to be analyzed:
Pearson's r (default), Spearman's rho, Kendall's TauA, Kendall's TauB,
Chatterjee's Xi (xiA: no assumption on ties, xiB: tie correction included).}

\item{alternative}{type of the alternative hypothesis concerning correlation
coefficient sign. Ignored if \code{type = 'bootstrap'}.}

\item{ci_type}{type of confidence intervals: BCA (default) or percentile.}

\item{conf_level}{confidence level used for computation of the confidence
intervals.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{n_iter}{number of the algorithm's iterations.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}
}
\value{
A numeric vector, matrix, or a data frame with the following elements:
numbers of pairwise complete observations, the requested correlation
coefficients in the entire data set, bootstrapped means and confidence
interval bounds, numbers of successful iterations of the algorithm,
numbers of re-samples in favor of the null hypothesis H0, numbers of
re-samples in favor of the alternative hypothesis H1, p values, and,
optionally p values adjusted for multiple testing.
}
\description{
Permutation and bootstrap tests for correlation which are not covered by
base R.
}
\details{
The permutation tests are done by permuting one of the correlation partners
(i.e. re-sampling without replacement). The bootstrap tests employ
re-sampling of data points (i.e. X and Y pairs) with replacement.
Please note that bootstrapping is likely going to break trends in data
investigated by the xi family of correlation coefficients: the bootstrapped
mean is likely to be substantially lower than the correlation coefficient
for the entire data set. Nevertheless, bootstrapped xi coefficients with
the respective confidence intervals are good estimate of interpolation of the
data trends.
Please note, that bootstrapped p values refer to the two-sided setting.
If a numeric matrix or data frame is provided as the \code{x} argument, the
correlations are performed in pairwise manner between columns of the input
object.
}
\references{
Chatterjee S. A New Coefficient of Correlation. J Am Stat Assoc (2021)
116:2009–2022. doi:10.1080/01621459.2020.1758115
}
