% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_correlation.R
\name{f_cov}
\alias{f_cov}
\alias{f_cor}
\alias{f_cov.default}
\alias{f_cov.list}
\alias{f_cov.matrix}
\alias{f_cov.data.frame}
\alias{f_cor.default}
\alias{f_cor.list}
\alias{f_cor.matrix}
\alias{f_cor.data.frame}
\title{Covariance and correlation coefficients.}
\usage{
f_cov(x, ...)

f_cor(x, ...)

\method{f_cov}{default}(x, y, method = c("pearson", "spearman"), ...)

\method{f_cov}{list}(x, method = c("pearson", "spearman"), ...)

\method{f_cov}{matrix}(x, method = c("pearson", "spearman"), ...)

\method{f_cov}{data.frame}(x, method = c("pearson", "spearman"), ...)

\method{f_cor}{default}(
  x,
  y,
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  ...
)

\method{f_cor}{list}(
  x,
  y,
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  ...
)

\method{f_cor}{matrix}(
  x,
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  ...
)

\method{f_cor}{data.frame}(
  x,
  method = c("pearson", "spearman", "kendallA", "kendallB", "xiA", "xiB"),
  ...
)
}
\arguments{
\item{x}{a numeric vector, a list of numeric vectors, matrix or a data frame.}

\item{...}{extra arguments passed to methods.}

\item{y}{a numeric vector.}

\item{method}{type of covariance or correlation to be analyzed:
Pearson's r (default), Spearman's rho, Kendall's TauA, Kendall's TauB,
Chatterjee's Xi (xiA: no assumption on ties, xiB: tie correction included).}
}
\value{
Methods for vectors and lists return always two-element numeric vector with
the number of complete observations and the requested coefficient.
}
\description{
Alternatives to base R \code{\link[stats]{cov}}
and \code{\link[stats]{cor}} provided by \code{f_cov()} and \code{f_cor()} function
family.
}
\details{
The functions always use pairwise complete observations. Except for
Kednall's tau, they are not faster than the usual R functions which are
implemented in C.
If a matrix or a data frame is provided as the input, the columns are
handled as variables. The correlation-calculating functions offer also the
novel correlation coefficient Xi proposed by Chatterjee. Note that the later
correlation coefficients are generated with random resolution of rank ties
and may hence vary from run to run. Furthermore, those correlation
coefficients are not symmetric.
}
\references{
Chatterjee S. A New Coefficient of Correlation. J Am Stat Assoc (2021)
116:2009–2022. doi:10.1080/01621459.2020.1758115
}
