% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{f_enrichment}
\alias{f_enrichment}
\alias{f_enrichment.default}
\title{Set enrichment with Fisher's exact test and random sampling.}
\usage{
f_enrichment(x, ...)

\method{f_enrichment}{default}(
  x,
  dict,
  all = NULL,
  type = c("fisher", "random"),
  laplace = 1,
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  adj_method = "none",
  .parallel = FALSE,
  .n_chunks = NULL,
  ...
)
}
\arguments{
\item{x}{a character vector for which enrichment will be investigated, e.g.
a vector of identifiers of significantly regulated genes.}

\item{...}{extra arguments, currently none specified.}

\item{dict}{dictionary of sets. A named list of character vectors, where each
list element represents a feature set.}

\item{all}{a character vector of all investigated features. If \code{NULL}, those
features will be inferred from \code{dict}.}

\item{type}{type of enrichment test as described in details. Fisher's exact
test by default. Note that the random sampling method is considerably slower.}

\item{laplace}{Laplace smoother.}

\item{ci_type}{type of confidence intervals: BCA (default) or percentile.
Ignored for Fisher's tests.}

\item{conf_level}{confidence level used for computation of the confidence
intervals. Ignored for Fisher's tests.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{n_iter}{number of the algorithm's iterations. Ignored for Fisher's
tests.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{.parallel}{logical, should the analysis be run in parallel? At the
moment, only the R-side parallelization via \code{\link[furrr]{future_map}}
is implemented. This option is available only for \code{type = 'random'}.}

\item{.n_chunks}{number of chunks of \code{dict} passed to parallel workers.
If \code{.n_chunks = NULL}, the number of chunks will be set to the umber of
available cores minus 1.}
}
\value{
a numeric matrix or a data frame with the following columns:
\itemize{
\item \code{entry_name}: name of the entry of the dictionary of sets, i.e. set name
(only for data frame output)
\item \code{n_entry}: numbers of all available features belonging to the set
\item \code{n_x_total}: the total number of elements in vector \code{x}
\item \code{n_intersect}: numbers of elements of \code{x} belonging to the set
\item \code{or}: odds ratio (OR)
\item \code{se_log_or}: approximate standard error of log(OR) computed as square
root of sum of inversions of cells of the contingency table (only for Fisher's
exact test).
\item \code{lower_ci} and \code{upper_ci}: lower and upper bound of confidence (only for
random sampling)
\item \code{iter_number}, \code{h0_number}, and \code{h1_number}: available only for the
random sampling algorithm; total number of successful iterations, numbers of
iterations in favor of the null hypothesis H0, and numbers of iterations in
favor of the alternative hypothesis H1
\item \code{p_value} and, optionally, \code{p_adjusted}: p values and p values adjusted for
multiple testing.
}
}
\description{
The function tests for enrichment of a vector with elements following some
classification system. A good example of enrichment analyses are gene
ontology term enrichment. The overlap between elements of a vector of
interest with elements of a particular set is investigated by two methods: by
Fisher's exact test and random sampling.
}
\details{
Fisher's exact test operates with a \eqn{2 \times 2} contingency table and
checks if an intersection of the vector \code{x} and the given set is larger than
an intersection of all available features and the set. The odds ratio (OR)
statistic generated by Fisher's exact test is defined as follows:
\deqn{OR = \frac{n_{x \in set} \times n_{all \notin set}}{n_{x \notin set} \times n_{all \in set}}}
where \eqn{n_{x \in set}} stands for the number of elements of vector \code{x}
belonging to the set,
\eqn{n_{x \notin set}} represents the number of elements of vector \code{x} not
belonging to the set,
\eqn{n_{all \in set}} is the number of all available features belonging to
the set, and
\eqn{n_{all \notin set}} is the number of all available features not
belonging to the set.
In other words, Fisher's exact test tests for \eqn{OR \neq 1}.

Random sampling test compares the number of elements of the particular set
in vector x with a large number of random samples draws from all available
features. Each of those samples is of equal size as vector \code{x}. The OR
statistic is defined as follows:
\deqn{OR = \frac{1}{n_{iter}} \times \sum^{n_{iter}}_i \frac{n_{x \in set}}{n_{sample,i \in set}}}
where \eqn{n_{iter}} is the number of algorithm iterations,
\eqn{n_{x \in set}} represents the number of observations in \code{x} belonging
to the set, and
\eqn{n_{sample,i \in set}} is the number of elements of the i^th random
sample shared with the set of interest.
The random sampling method allows for estimation of confidence intervals of
OR either with the bias-corrected accelerated (BCA) method or a simple 95^th
percentile range.

Independently of the enrichment test type, zero-division may occur during
calculation of OR. To avoid it, the function allows for specifying of a
\href{https://en.wikipedia.org/wiki/Additive_smoothing}{Laplace smoother}.
}
