% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman_tests.R
\name{f_friedman_test}
\alias{f_friedman_test}
\alias{f_friedman_test.default}
\alias{f_friedman_test.matrix}
\alias{f_friedman_test.data.frame}
\title{Friedman test.}
\usage{
f_friedman_test(x, ...)

\method{f_friedman_test}{default}(x, f, b, as_data_frame = FALSE, ...)

\method{f_friedman_test}{matrix}(
  x,
  f,
  b,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_friedman_test}{data.frame}(
  x,
  f,
  b,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector, list of numeric vectors, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{f}{a factor or integer vector used for splitting by the treatment.}

\item{b}{a factor of integer vector which defines the blocks.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
A numeric vector, matrix, or data frame with the following information:
total observation number (\code{n}), number of treatment groups (\code{k}), number of
blocks (\code{b}), test statistic Q, tiee correction factor, degrees of freedom,
p-value, Kendall's W effect size statistic, and, optionally, p value adjusted
for multiple testing.
}
\description{
Faster and more flexible counterparts of base R's
\code{\link[stats]{friedman.test}}.
}
\details{
The matrix and data frame methods perform a series of Friedman tests: one for
each column. Kednall's W serves as an effect size statistic.
}
