% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{f_kappa}
\alias{f_kappa}
\alias{f_kappa.default}
\alias{f_kappa.matrix}
\alias{f_kappa.data.frame}
\title{Cohen's kappa inter-rater reliability statistic.}
\usage{
f_kappa(x, ...)

\method{f_kappa}{default}(
  x,
  y,
  method = c("unweighted", "equal", "fleiss"),
  as_data_frame = FALSE,
  ...
)

\method{f_kappa}{matrix}(
  x,
  y = NULL,
  method = c("unweighted", "equal", "fleiss"),
  as_data_frame = FALSE,
  ...
)

\method{f_kappa}{data.frame}(
  x,
  y = NULL,
  method = c("unweighted", "equal", "fleiss"),
  as_data_frame = FALSE,
  ...
)
}
\arguments{
\item{x}{a factor or integer vector, an integer matrix, or a data frame
with factor or integer columns.}

\item{...}{additional arguments passed to methods.}

\item{y}{\code{NULL} or a factor or integer vector, an integer matrix, or
a data frame with factor or integer columns.}

\item{method}{weighting method: \code{"unweighted"} (default) returns unweighted kappa,
\code{"equal"} computes kappas with equally spaced weighting, and \code{"fleiss"}
calculates kappas with Fleiss - Cohen weights.}

\item{as_data_frame}{logical: should the output be coerced to a data frame?}
}
\value{
The default \code{f_kappa()} method for a pair of vectors returns a numeric vector
with the number of complete observations and kappa value.
The method for a pair of matrices or data frames returns
a numeric matrix or a data frame with the numbers of complete observations and
kappa values.
If only one matrix or data frame is provided, i.e. \code{y = NULL}, kappa
coefficients are computed for all pairs of columns in \code{x}.
If \code{as_data_frame = TRUE}, the output is coerced to a data frame, and,
optionally appended with names of variables in matrices or data frames.
Note: because of checks of level compatibility, the function is way faster
for integers than factors.
}
\description{
The function computes unweighted Cohen's kappa, or Cohen's kappa weighted
with equally spaced or Fleiss - Cohen weights.
}
\details{
If \code{x} and \code{y} are matrices, the kappas are computed for the corresponding
columns.
Hence, these matrices or data frames need to have equal dimensions.
\code{NA} values are silently removed.
}
\references{
Cohen J. A Coefficient of Agreement for Nominal Scales.
Educ Psychol Meas (1960) 20:37–46. doi:10.1177/001316446002000104

Fleiss JL, Cohen J, Everitt BS. Large sample standard errors of kappa and
weighted kappa. Psychol Bull (1969) 72:323–327. doi:10.1037/h0028106
}
