% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_tests.R
\name{f_kappa_test}
\alias{f_kappa_test}
\alias{f_kappa_test.default}
\alias{f_kappa_test.matrix}
\alias{f_kappa_test.data.frame}
\title{Cohen's kappa inter-rater reliability statistic: permutation and bootstrap
tests}
\usage{
f_kappa_test(x, ...)

\method{f_kappa_test}{default}(
  x,
  y,
  type = c("permutation", "bootstrap"),
  method = c("unweighted", "equal", "fleiss"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  ...
)

\method{f_kappa_test}{matrix}(
  x,
  y,
  type = c("permutation", "bootstrap"),
  method = c("unweighted", "equal", "fleiss"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  adj_method = "none",
  ...
)

\method{f_kappa_test}{data.frame}(
  x,
  y,
  type = c("permutation", "bootstrap"),
  method = c("unweighted", "equal", "fleiss"),
  alternative = c("two.sided", "less", "greater"),
  ci_type = c("bca", "percentile"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  n_iter = 1000,
  adj_method = "none",
  ...
)
}
\arguments{
\item{x}{a factor or integer vector, an integer matrix, or a data frame
with factor or integer columns.}

\item{...}{additional arguments passed to methods.}

\item{y}{a factor or integer vector, an integer matrix, or a data frame
with factor or integer columns.}

\item{type}{type of the test: permutation (default) or bootstrap.}

\item{method}{weighting method: \code{"unweighted"} (default) returns unweighted kappa,
\code{"equal"} computes kappas with equally spaced weighting, and \code{"fleiss"}
calculates kappas with Fleiss - Cohen weights.}

\item{alternative}{type of the alternative hypothesis concerning correlation
coefficient sign. Ignored if \code{type = 'bootstrap'}.}

\item{ci_type}{type of confidence intervals: BCA (default) or percentile.}

\item{conf_level}{confidence level used for computation of the confidence
intervals.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{n_iter}{number of the algorithm's iterations.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}
}
\value{
A numeric vector, matrix, or a data frame with the following elements:
numbers of pairwise complete observations, the requested correlation
coefficients in the entire data set, bootstrapped means and confidence
interval bounds, numbers of successful iterations of the algorithm,
numbers of re-samples in favor of the null hypothesis H0, numbers of
re-samples in favor of the alternative hypothesis H1, p values, and,
optionally p values adjusted for multiple testing.
}
\description{
The function computes permuted and bootstrapped unweighted Cohen's kappa,
or Cohen's kappa weighted with equally spaced or Fleiss - Cohen weights.
}
\details{
If \code{x} and \code{y} are matrices, the kappas are computed for the corresponding
columns.
Hence, these matrices or data frames need to have equal dimensions.
\code{NA} values are silently removed.
}
\references{
Cohen J. A Coefficient of Agreement for Nominal Scales.
Educ Psychol Meas (1960) 20:37–46. doi:10.1177/001316446002000104

Fleiss JL, Cohen J, Everitt BS. Large sample standard errors of kappa and
weighted kappa. Psychol Bull (1969) 72:323–327. doi:10.1037/h0028106
}
