% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_tests.R
\name{f_kruskal_test}
\alias{f_kruskal_test}
\alias{f_kruskal_test.default}
\alias{f_kruskal_test.list}
\alias{f_kruskal_test.matrix}
\alias{f_kruskal_test.data.frame}
\title{Kruskal-Wallis test.}
\usage{
f_kruskal_test(x, ...)

\method{f_kruskal_test}{default}(x, f, as_data_frame = FALSE, ...)

\method{f_kruskal_test}{list}(x, as_data_frame = FALSE, ...)

\method{f_kruskal_test}{matrix}(
  x,
  f,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_kruskal_test}{data.frame}(
  x,
  f,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric vector, a list of numeric vectors, numeric matrix, or a
numeric data frame.}

\item{...}{extra arguments passed to methods.}

\item{f}{a factor or an integer vector used for splitting.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
A numeric vector, numeric matrix, or a numeric data frame with the following
information: number of complete observations, number of groups,
tie-corrected test statistic H, tie correction factor, degrees of freedom,
p-value, eta-square effect size statistic, and, optionally, p-value
adjusted for multiple testing.
}
\description{
A faster and more versatile alternative to the base R's
\code{\link[stats]{kruskal.test}}.
}
\details{
If a numeric matrix or a numeric data frame is provided as the argument,
a series of Kruskal-Wallis tests is performed: one for each column.
The test statistic \eqn{H} is
\href{https://en.wikipedia.org/wiki/Kruskal\%E2\%80\%93Wallis_test}{corrected for ties},
and the p-value is derived from chi-square distribution.
The eta-square effect size statistic of computed with the following formula:
\deqn{\eta^2 = \frac{H - k + 1}{N -k}}
where \eqn{k} represents the number
of groups, and \eqn{N} is the number of observations.
}
