% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_tests.R
\name{f_ks_test}
\alias{f_ks_test}
\alias{f_ks_test.default}
\alias{f_ks_test.matrix}
\alias{f_ks_test.data.frame}
\title{Two-sample Kolmogorov-Smirnov tests}
\usage{
f_ks_test(x, ...)

\method{f_ks_test}{default}(
  x,
  f,
  alternative = c("two.sided", "greater", "less"),
  as_data_frame = FALSE,
  ...
)

\method{f_ks_test}{matrix}(
  x,
  f,
  alternative = c("two.sided", "greater", "less"),
  as_data_frame = FALSE,
  adj_method = "none",
  ...
)

\method{f_ks_test}{data.frame}(
  x,
  f,
  alternative = c("two.sided", "greater", "less"),
  as_data_frame = FALSE,
  adj_method = "none",
  ...
)
}
\arguments{
\item{x}{a list of two numeric vectors, vector, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{f}{a factor or integer vector used for splitting. If it contains more
than two non-empty levels, the first two will be used in the analysis.}

\item{alternative}{type of the alternative hypothesis concerning difference
of means.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}
}
\value{
A numeric vector, a matrix, or a data frame (if \code{as_data_frame = TRUE}) with
numbers of complete observations in the first and second sample, the testing
statistic d, and asymptotic (approximate) p value.
If \code{as_data_frame = TRUE}, the output data frame is appended with names of
variables.
In matrices and data frames, the testing is done column-wise.
}
\description{
Rcpp implementations of the two-sample Kolmogorov-Smirnov test.
}
