% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_tests.R
\name{f_levene_test}
\alias{f_levene_test}
\alias{f_levene_test.list}
\alias{f_levene_test.default}
\alias{f_levene_test.matrix}
\alias{f_levene_test.data.frame}
\title{Levene test for equality of variances.}
\usage{
f_levene_test(x, ...)

\method{f_levene_test}{list}(x, type = c("standard", "bf"), as_data_frame = FALSE, ...)

\method{f_levene_test}{default}(x, f, type = c("standard", "bf"), as_data_frame = FALSE, ...)

\method{f_levene_test}{matrix}(
  x,
  f,
  type = c("standard", "bf"),
  as_data_frame = FALSE,
  safely = FALSE,
  ...
)

\method{f_levene_test}{data.frame}(
  x,
  f,
  type = c("standard", "bf"),
  as_data_frame = FALSE,
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector, list of numeric vectors, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{type}{type of the test: standard which employs group mean as a
centrality measure, or Brown-Forsythe, which operates with group medians.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{f}{a factor or integer vector used for splitting.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
a numeric vector, matrix, or a data frame with the following
information: number of complete observations, number of groups, the test
statistic F, number of degrees of freedom (\code{df1}: nominator, \code{df2}
denominator) and p value.
}
\description{
Levene test checks for equality of variances of a set of samples (H0: all
variances equal, H1: at least one variance not equal as compared with the
rest). The family of \code{f_levene_test} functions provide faster and more
versatile alternatives to \code{\link[car]{leveneTest}}.
}
\details{
In a numeric matrix or data frame is provided as an argument, a series of
Levene tests with the columns is performed. The argument \code{type} allow for
choice of the group's centrality statistic: mean (\code{type = 'standard'}) or
median (\code{type = 'bf'}). This later option results in a Brown - Forsythe test
and may be a better option for strongly skewed distributions.
}
