% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_estimates.R
\name{f_meta}
\alias{f_meta}
\alias{f_meta.default}
\alias{f_meta.list}
\alias{f_meta.matrix}
\alias{f_meta.data.frame}
\title{Inverse variance-weighted meta-estimates.}
\usage{
f_meta(y, e, ...)

\method{f_meta}{default}(
  y,
  e,
  type = c("fixed", "random"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  ...
)

\method{f_meta}{list}(
  y,
  e,
  type = c("fixed", "random"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_meta}{matrix}(
  y,
  e,
  type = c("fixed", "random"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_meta}{data.frame}(
  y,
  e,
  type = c("fixed", "random"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{y}{a numeric vector, list, matrix, or a data frame with estimates
obtained in single studies.}

\item{e}{a numeric vector, list, matrix, or a data frame with errors, e.g.
standard errors or other errors at the same scale as \code{y}.}

\item{...}{extra arguments passed to methods.}

\item{type}{algorithm used for calculation of the meta-estimates. \code{fixed}
(default) assumes that there's no significant between-study variance.
\code{random} estimates the between-study variance (i.e. tau-square) with the
DerSimonian-Lair algorithm.}

\item{alternative}{type of the alternative hypothesis concerning difference
of means.}

\item{conf_level}{confidence level used for computation of the confidence
intervals.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
A numeric vector, matrix, or data frame with the following statistics:
number of studies, Cochran's Q statistic of between-study variance,
p value for significant between-study variance determined from chi-square
distribution of Q, tau-square estimate of between-study variance,
meta-estimate calculated by the inverse-variance method, standard error of
the meta-estimate (SEM), lower and upper bound of the confidence interval,
test statistic Z, and p-value for significant meta-effect.
}
\description{
The functions compute inverse variance-weighted meta-estimates of an effect
with the fixed and random between-study variance (DerSimonian-Lair algorithm).
While those tools are well suited for high throughput analyses of e.g.
transcriptomic data, the excellent R package \code{meta} with richer algorithm
and quality control options is recommended for  meta-analyses of clinical
studies.
}
\details{
In case of lists, matrices or data frames provided as the function arguments,
the meta-estimates are computed in an element- and column-wise manner.
This means that each element or each column is treated as a a vector of
estimates or errors obtained in a single study.
}
\references{
Schwarzer G, Carpenter JR, Rücker G. Meta-Analysis with R. (2015)
doi:10.1007/978-3-319-21416-0
}
