% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_anova.R
\name{f_one_anova}
\alias{f_one_anova}
\alias{f_one_anova.list}
\alias{f_one_anova.default}
\alias{f_one_anova.matrix}
\alias{f_one_anova.data.frame}
\title{One-way analysis of variance (ANOVA).}
\usage{
f_one_anova(x, ...)

\method{f_one_anova}{list}(x, as_data_frame = FALSE, ...)

\method{f_one_anova}{default}(x, f, as_data_frame = FALSE, ...)

\method{f_one_anova}{matrix}(
  x,
  f,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_one_anova}{data.frame}(
  x,
  f,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector, list of numeric vectors, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{f}{a factor or integer vector used for splitting.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
A numeric vector, matrix, or a data frame containing the following
information: number of complete observations, number of groups,
sum-of-squares: between the groups, within the groups and total, test
statistic F, degrees of freedom (nominator: \code{df1}, denominator \code{df2}),
p value, eta-square effect size statistic, and, optionally, multiple
testing-adjusted p value.
}
\description{
A faster and more versatile version of one-way ANOVA genuinely implemented
by \code{\link[stats]{aov}}.
}
\details{
If a numeric matrix or a numeric data frame is provided as the first
argument, a series of one-way ANOVAs is performed: one for each column.
}
