% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_block_anova.R
\name{f_one_block_anova}
\alias{f_one_block_anova}
\alias{f_one_block_anova.default}
\alias{f_one_block_anova.matrix}
\alias{f_one_block_anova.data.frame}
\title{One-way ANOVA for block designs.}
\usage{
f_one_block_anova(x, ...)

\method{f_one_block_anova}{default}(x, f, b, as_data_frame = FALSE, ...)

\method{f_one_block_anova}{matrix}(
  x,
  f,
  b,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_one_block_anova}{data.frame}(
  x,
  f,
  b,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector, list of numeric vectors, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{f}{a factor or integer vector used for splitting by the treatment.}

\item{b}{a factor of integer vector which defines the blocks.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
A numeric vector, numeric matrix, or a numeric data frame with the following
information:
\itemize{
\item \code{n}, \code{k}, \code{b}: numbers of complete observations, treatment groups, and
blocks, respectively
\item \code{ss_treatment}, \code{ss_block}, \code{ss_within}, \code{ss_total}: sum of squares, for,
respectively treatment effect, blocks, errors (i.e. within
block and treatment groups), and all observations
\item \code{f}, \code{df1}, \code{df2}, \code{p}, \code{etasq}: test statistic F, nominator and
denominator degrees of freedom, p-value, and eta-square effect size statistic
for the treatment and block effect
\item \code{p_adjusted}: optional, multiple testing-adjusted p values for the
treatment and block effect. The adjustment is done for matrix and data frame
inputs in a column-wise manner
}
}
\description{
The \code{f_one_block_anova} family allows for computation of onw-way analysis
of variance (ANOVA) with effects of a single treatment factor \code{f} and
partitioning of error between blocks defined by a single grouping factor \code{b}.
}
\details{
As such, the results are equivalent to the traditional R's ANOVA build with
the \code{x ~ f  + Error(b)} formula. If your data has a more sophisticated block
design, you are better served with the base R's function
\code{\link[stats]{aov}}.
If a numeric matrix or a numeric data frame is provided ast the function
argument, a series of ANOVAs is computed in a column-wise manner.
}
