% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_tests.R
\name{f_shapiro_test}
\alias{f_shapiro_test}
\alias{f_shapiro_test.default}
\alias{f_shapiro_test.list}
\alias{f_shapiro_test.matrix}
\alias{f_shapiro_test.data.frame}
\title{Shapiro-Wilk tests.}
\usage{
f_shapiro_test(x, ...)

\method{f_shapiro_test}{default}(x, f = NULL, as_data_frame = FALSE, ...)

\method{f_shapiro_test}{list}(x, as_data_frame = FALSE, ...)

\method{f_shapiro_test}{matrix}(x, f = NULL, as_data_frame = FALSE, ...)

\method{f_shapiro_test}{data.frame}(x, f = NULL, as_data_frame = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector, list of numeric vectors, matrix or a data frame.}

\item{...}{extra arguments passed to methods.}

\item{f}{an optional factor or integer vector used for splitting.}

\item{as_data_frame}{should the output be formatted as a data frame?}
}
\value{
If the splitting vector \code{f = NULL}, the functions return numeric vectors,
matrices or data frames. Otherwise, a list of those objects are returned,
where each element represents a single non-empty level of \code{f}. The output
contains information on the number of complete cases, test statistic W, and p
values.
}
\description{
The functions compute Shapiro-Wilk tests for various inputs: single vectors,
a list of vectors, numeric matrices and data frames.
}
\details{
Because the base R's \code{\link[stats]{shapiro.test}} is already a fast C
implementation, we decided not to include an extra C++ version: functions
of the \code{f_shapiro_test} are just simple R wrappers designed for various input
objects.
If a data frame or a matric is provided as the \code{x} argument, a series of
Shapiro-Wilk tests is performed, where columns are treated as single
variables.
}
