% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_split}
\alias{f_split}
\alias{f_split.default}
\alias{f_split.matrix}
\alias{f_split.data.frame}
\title{Sub-setting of a vector or a matrix by a splitting factor.}
\usage{
f_split(x, f, ...)

\method{f_split}{default}(x, f, ...)

\method{f_split}{matrix}(x, f, ...)

\method{f_split}{data.frame}(x, f, ...)
}
\arguments{
\item{x}{a numeric vector, matrix, or data frame.}

\item{f}{a factor or integer vector used for splitting.}

\item{...}{extra arguments passed to methods.}
}
\value{
a list of vectors, matrices, or data frames.
}
\description{
Alternatives to base R \code{split()} for numeric vectors, matrices,
and data frames.
}
