% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{f_t_test}
\alias{f_t_test}
\alias{f_t_test.list}
\alias{f_t_test.default}
\alias{f_t_test.matrix}
\alias{f_t_test.data.frame}
\title{Student's, Welch's, and paired T test.}
\usage{
f_t_test(x, ...)

\method{f_t_test}{list}(
  x,
  type = c("standard", "welch", "paired"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  ...
)

\method{f_t_test}{default}(
  x,
  f,
  type = c("standard", "welch", "paired"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  ...
)

\method{f_t_test}{matrix}(
  x,
  f,
  type = c("standard", "welch", "paired"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)

\method{f_t_test}{data.frame}(
  x,
  f,
  type = c("standard", "welch", "paired"),
  alternative = c("two.sided", "less", "greater"),
  conf_level = 0.95,
  as_data_frame = FALSE,
  adj_method = "none",
  safely = FALSE,
  ...
)
}
\arguments{
\item{x}{a list of two numeric vectors, vector, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{type}{type of the test: standard/Student's, Welch's, or paired
T test.}

\item{alternative}{type of the alternative hypothesis concerning difference
of means.}

\item{conf_level}{confidence level used for computation of the confidence
intervals.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{f}{a factor or integer vector used for splitting. If it contains more
than two non-empty levels, the first two will be used in the analysis.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
a vector/matrix or a data frame with numbers of complete
observations in the investigated groups, T statistics, degrees of freedom,
p values, means in the first group, means in the second group, estimate aka
difference of means (the second groups versus the first one, or, in case of
paired tests, mean of differences), confidence intervals of the difference,
pooled variance used for effect size estimation, and Cohen's d effect size
statistic.
Optionally, multiple-testing-adjusted p values are returned in the last
column.
}
\description{
A faster and more versatile alternative to the standard R
\code{\link[stats]{t.test}}.
}
\details{
In case of matrices or data frames, a series of T tests is performed
in a column-wise manner, i.e. each column is treated as a separate variable.
}
