% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_table}
\alias{f_table}
\alias{f_table.default}
\title{Contingency tables.}
\usage{
f_table(x, f, ...)

\method{f_table}{default}(x, f = NULL, as_table = TRUE, ...)
}
\arguments{
\item{x}{a vector, matrix, or data frame: numeric or factor,}

\item{f}{a factor or integer vector used for splitting, if \code{NULL}, single
dimension table is returned.}

\item{...}{extra arguments passed to methods.}

\item{as_table}{should an instance of \code{table} class be returned?}
}
\value{
a numeric matrix with counts.
}
\description{
Alternative to base R's \code{table()}.
}
