% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_tests.R
\name{f_wilcox_test}
\alias{f_wilcox_test}
\alias{f_wilcox_test.list}
\alias{f_wilcox_test.default}
\alias{f_wilcox_test.matrix}
\alias{f_wilcox_test.data.frame}
\title{Wilcoxon or Mann-Whitney test.}
\usage{
f_wilcox_test(x, ...)

\method{f_wilcox_test}{list}(
  x,
  type = c("standard", "paired"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  correct = TRUE,
  hl_estimate = FALSE,
  as_data_frame = FALSE,
  ...
)

\method{f_wilcox_test}{default}(
  x,
  f,
  type = c("standard", "paired"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  correct = TRUE,
  hl_estimate = FALSE,
  as_data_frame = FALSE,
  ...
)

\method{f_wilcox_test}{matrix}(
  x,
  f,
  type = c("standard", "paired"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  correct = TRUE,
  hl_estimate = FALSE,
  adj_method = "none",
  safely = FALSE,
  as_data_frame = FALSE,
  ...
)

\method{f_wilcox_test}{data.frame}(
  x,
  f,
  type = c("standard", "paired"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  correct = TRUE,
  hl_estimate = FALSE,
  adj_method = "none",
  safely = FALSE,
  as_data_frame = FALSE,
  ...
)
}
\arguments{
\item{x}{a list of two numeric vectors, vector, matrix or a data frame.
Vectors, matrices, and data frames have to contain numeric values.}

\item{...}{extra arguments passed to methods.}

\item{type}{type of the test: standard Wilcoxon rank test or paired Wilcoxon
signed rank test.}

\item{alternative}{type of the alternative hypothesis concerning difference
between the groups (group 2 against group 1).}

\item{exact}{logical, should p values be derived from Wilcoxon
distributions? Otherwise, p values will be computed with a normal
distribution approximation. Note that calculation of the exact p value is
possible only if there are no ties in ranks of the groups. If there are any
ties, normal distribution approximation will be used.}

\item{correct}{logical, should continuity correction be applied to the test
statistic?}

\item{hl_estimate}{logical, should Hodges-Lehman estimate of location be
computed? This option is computationally intensive, as default
(\code{hl_estimate = FALSE}), difference of medians of the analysis groups is
returned.}

\item{as_data_frame}{should the output be formatted as a data frame? This may
render the computation slower.}

\item{f}{a factor or integer vector used for splitting. If it contains more
than two non-empty levels, the first two will be used in the analysis.}

\item{adj_method}{multiple testing adjustment method, as specified for
\code{\link[stats]{p.adjust}}.}

\item{safely}{if \code{TRUE}, most execution errors will be turned into warnings
and \code{NA} values will be returned as the testing results. This option is
particularly useful for large analysis pipelines with possible data quality
issues.}
}
\value{
a vector, matrix, or a data frame with the following information;
numbers of samples in the first and the second group, Wilcoxon/Mann-Whitney
test statistic U (also known as V for the signed rank test), p value,
total number of observation pairs, fractions of pairs in favor of the NULL
hypothesis H0, fraction of pairs in favor of the alternative hypothesis H1,
estimate of difference between the groups (Hodges-Lehman location parameter
or difference of medians, group 2 against group 1, as defined by levels of
the splitting factor \code{f}),
biserial rank correlation coefficient r as an effect size metric, and,
optionally, multiple testing-adjusted p value.
}
\description{
A faster and more versatile alternative to R
\code{\link[stats]{wilcox.test}} performing Wilcoxon rank tests
(also known as Mann-Whitney test) and Wilcoxon signed rank test for paired
data.
}
\details{
In case of matrices or data frames, a series of Wilcoxon tests is performed
in a column-wise manner, i.e. each column is treated as a separate variable.
}
